Adds __service_name__ labels to improve data locality
