+++ pkg/phlaredb/head.go
@@ -6,7 +6,6 @@ import (
 	"os"
 	"path/filepath"
 	"sort"
-	"strings"
 	"sync"
 	"time"
 
@@ -417,51 +416,6 @@ func (h *Head) Ingest(ctx context.Context, p *profilev1.Profile, id uuid.UUID, e
 	h.metaLock.Unlock()
 
 	return nil
-}
-
-func labelsForProfile(p *profilev1.Profile, externalLabels ...*typesv1.LabelPair) ([]phlaremodel.Labels, []model.Fingerprint) {
-	// build label set per sample type before references are rewritten
-	var (
-		sb                                             strings.Builder
-		lbls                                           = phlaremodel.NewLabelsBuilder(externalLabels)
-		sampleType, sampleUnit, periodType, periodUnit string
-		metricName                                     = phlaremodel.Labels(externalLabels).Get(model.MetricNameLabel)
-	)
-
-	// set common labels
-	if p.PeriodType != nil {
-		periodType = p.StringTable[p.PeriodType.Type]
-		lbls.Set(phlaremodel.LabelNamePeriodType, periodType)
-		periodUnit = p.StringTable[p.PeriodType.Unit]
-		lbls.Set(phlaremodel.LabelNamePeriodUnit, periodUnit)
-	}
-
-	profilesLabels := make([]phlaremodel.Labels, len(p.SampleType))
-	seriesRefs := make([]model.Fingerprint, len(p.SampleType))
-	for pos := range p.SampleType {
-		sampleType = p.StringTable[p.SampleType[pos].Type]
-		lbls.Set(phlaremodel.LabelNameType, sampleType)
-		sampleUnit = p.StringTable[p.SampleType[pos].Unit]
-		lbls.Set(phlaremodel.LabelNameUnit, sampleUnit)
-
-		sb.Reset()
-		_, _ = sb.WriteString(metricName)
-		_, _ = sb.WriteRune(':')
-		_, _ = sb.WriteString(sampleType)
-		_, _ = sb.WriteRune(':')
-		_, _ = sb.WriteString(sampleUnit)
-		_, _ = sb.WriteRune(':')
-		_, _ = sb.WriteString(periodType)
-		_, _ = sb.WriteRune(':')
-		_, _ = sb.WriteString(periodUnit)
-		t := sb.String()
-		lbls.Set(phlaremodel.LabelNameProfileType, t)
-		lbs := lbls.Labels().Clone()
-		profilesLabels[pos] = lbs
-		seriesRefs[pos] = model.Fingerprint(lbs.Hash())
-
-	}
-	return profilesLabels, seriesRefs
 }
 
 // LabelValues returns the possible label values for a given label name.
